// apps/fleet/fleet-mfe/src/app/fleet-product/line/shared/models/address.model.ts
export class Address {
  addressId: string;
  type: string;
  houseNumber: string;
  boxNumber: string;
  subHouseNumber: string;
  street: string;
  streetId: string;
  postalCode: string;
  municipality: string;
  country?: string;
  countryCode?: string;
  floorNumber?: string;
  addressType?: string;
  tinaLocationId?: string;
}